requirements.txt
asgiref==3.2.10
Django==3.0.5
django-filter==2.3.0
djangorestframework==3.11.0
djongo==1.3.3
importlib-metadata==1.7.0
Markdown==3.2.2
numpy==1.19.0
pandas==1.0.5
pymongo==3.10.1
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
sqlparse==0.2.4
zipp==3.1.0
